adding some tweaks, adding icons, adding outline material
