+++ ArsonistMod/ArsonistMod.manifest
@@ -1,5 +1,5 @@
 ManifestFileVersion: 0
-CRC: 3211240849
+CRC: 3560685121
 AssetBundleManifest:
   AssetBundleInfos:
     Info_0:
