+++ ArsonistMod/Content/Arsonist.cs
@@ -486,7 +486,7 @@ namespace ArsonistMod.Modules.Survivors
                 skillName = prefix + "PRIMARY_FIRESPRAY_SCEPTER_NAME",
                 skillNameToken = prefix + "PRIMARY_FIRESPRAY_SCEPTER_NAME",
                 skillDescriptionToken = prefix + "PRIMARY_FIRESPRAY_SCEPTER_DESCRIPTION",
-                skillIcon = Modules.AssetsArsonist.mainAssetBundle.LoadAsset<Sprite>("texPrimaryIcon"),
+                skillIcon = Modules.AssetsArsonist.mainAssetBundle.LoadAsset<Sprite>("scepterFireballIcon"),
                 activationState = new EntityStates.SerializableEntityStateType(typeof(SkillStates.FireSprayScepter)),
                 activationStateMachineName = "Weapon",
                 baseMaxStock = 1,
@@ -511,7 +511,7 @@ namespace ArsonistMod.Modules.Survivors
                 skillName = prefix + "PRIMARY_FLAMETHROWER_SCEPTER_NAME",
                 skillNameToken = prefix + "PRIMARY_FLAMETHROWER_SCEPTER_NAME",
                 skillDescriptionToken = prefix + "PRIMARY_FLAMETHROWER_SCEPTER_DESCRIPTION",
-                skillIcon = Modules.AssetsArsonist.mainAssetBundle.LoadAsset<Sprite>("arsonistFlamethrower"),
+                skillIcon = Modules.AssetsArsonist.mainAssetBundle.LoadAsset<Sprite>("scepterFlamethrowerIcon"),
                 activationState = new EntityStates.SerializableEntityStateType(typeof(SkillStates.FlamethrowerScepter)),
                 activationStateMachineName = "Weapon",
                 baseMaxStock = 1,
@@ -696,8 +696,8 @@ namespace ArsonistMod.Modules.Survivors
                 skillName = prefix + "SPECIAL_MASOCHISM_SURGE_NAME",
                 skillNameToken = prefix + "SPECIAL_MASOCHISM_SURGE_NAME",
                 skillDescriptionToken = prefix + "SPECIAL_MASOCHISM_SURGE_DESCRIPTION",
-                skillIcon = Modules.AssetsArsonist.mainAssetBundle.LoadAsset<Sprite>("texSpecialIcon"),
-                activationState = new EntityStates.SerializableEntityStateType(typeof(SkillStates.MasochismSurge)),
+                skillIcon = Modules.AssetsArsonist.mainAssetBundle.LoadAsset<Sprite>("spiteIcon"),
+                activationState = new EntityStates.SerializableEntityStateType(typeof(SkillStates.Spite)),
                 activationStateMachineName = "Weapon",
                 baseMaxStock = 1,
                 baseRechargeInterval = 0f,
