+++ ArsonistUnityProject/Assets/ArsonistAssets/Materials/OutlineSurge.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc24200742f4c4dc09a2cb11ddce3aa6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
