--- ArsonistUnityProject/Assets/ArsonistAssets/VFX/Materials/FresnelSurge.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aafae51709f5c4f00b4a12647a3b7720
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
