+++ ArsonistWWISE/ArsonistWWISE.ethan.validationcache
@@ -20,7 +20,7 @@
 	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Game Parameters\Factory SoundSeed Air Game Syncs.wwu" DateLow="1725081864" DateHigh="0"/>
 	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Game Parameters\Factory Motion.wwu" DateLow="1725081864" DateHigh="0"/>
 	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Game Parameters\Default Work Unit.wwu" DateLow="1729666136" DateHigh="0"/>
-	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Events\Default Work Unit.wwu" DateLow="1729666136" DateHigh="0"/>
+	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Events\Default Work Unit.wwu" DateLow="1730121874" DateHigh="0"/>
 	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Effects\Factory Reflect.wwu" DateLow="1725081864" DateHigh="0"/>
 	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Effects\Factory McDSP Effects.wwu" DateLow="1725081864" DateHigh="0"/>
 	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Effects\Factory Mastering Suite.wwu" DateLow="1725081864" DateHigh="0"/>
@@ -41,6 +41,6 @@
 	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Audio Devices\Default Work Unit.wwu" DateLow="1725083132" DateHigh="0"/>
 	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Attenuations\Default Work Unit.wwu" DateLow="1725081864" DateHigh="0"/>
 	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\ArsonistWWISE.wproj" DateLow="1725081864" DateHigh="0"/>
-	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\ArsonistWWISE.ethan.wsettings" DateLow="1729666190" DateHigh="0"/>
-	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Actor-Mixer Hierarchy\Default Work Unit.wwu" DateLow="1729666155" DateHigh="0"/>
+	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\ArsonistWWISE.ethan.wsettings" DateLow="1730122194" DateHigh="0"/>
+	<ValidationCacheItem File="F:\coding\Arsonist-RoR2\ArsonistWWISE\Actor-Mixer Hierarchy\Default Work Unit.wwu" DateLow="1730121927" DateHigh="0"/>
 </ValidationCache>
