+++ ArsonistWWISE/ArsonistWWISE.ethan.wsettings
@@ -138,7 +138,7 @@
 				<Property Name="ObjectTabPanel\x64_CurrentObjects_{FB6925F8-3EDD-42F1-B446-654A30D756CF}" Type="string" Value=""/>
 				<Property Name="ProjectExplorer\AudioSelectionChannel1\ProjectExplorer2LastScrollState" Type="int32" Value="19"/>
 				<Property Name="ProjectExplorer\AudioSelectionChannel1\x64_ProjectExplorer2LastExpandedState" Type="string" Value="iC8JJ1oiVBFnJcijTz1EguraiCN1l8ZGou3FIio5d1b4u6dewGRQSoghoH6L4h1ocev2AiV9hEaWq82ubOW2CrsAkUoVOztOqMsVnEZ7HsOxCn6gH3RZRZnm0I541HfqAZbf579crUW29sX62WSJeWbT/kkwtahCmTuSSHOqtwc="/>
-				<Property Name="ProjectExplorer\AudioSelectionChannel2\ProjectExplorer2LastScrollState" Type="int32" Value="42"/>
+				<Property Name="ProjectExplorer\AudioSelectionChannel2\ProjectExplorer2LastScrollState" Type="int32" Value="37"/>
 				<Property Name="ProjectExplorer\AudioSelectionChannel2\x64_ProjectExplorer2LastExpandedState" Type="string" Value="iC8JJ1oiVBFnJcijTz1Ego7YomTtQD9Ah3KzM9XBAUbq2ogjdZfGRqLtxSIqOXdWQYkKzzgwj0m57gWppK02Rfi7p17AZFBKiCGgfoviHWhx6/YCJX2ERparza5s5bYKuwCRShU7O06oyxWcRnsew7EKfqAfdFlFmebQjnjUd+oBlt/nv1ytRbb2xfrZZIl5Fv+qmFb8ckWZLY0UbVZ06GbT/kkwtahCmTuSSHOqtwc="/>
 				<Property Name="ProjectExplorer\EventsSelectionChannel1\ProjectExplorer2LastScrollState" Type="int32" Value="0"/>
 				<Property Name="ProjectExplorer\EventsSelectionChannel1\x64_ProjectExplorer2LastExpandedState" Type="string" Value="/PCbpkrYfEir3Vb3HLKhJonNaFkjwCdPrfnx4vk52xHe9ZowHwPgTJubwKU2J1ND"/>
